Add option to force match images height in a specific row
This allows the user to force the height of all images in a row
to be the same. This is used when a user wants to display several
images side by side with the same ratio: because of the way the
grid handles outer gutters, the width of the
far left and far right images are always greater than the inside
images (by one gutter).
The solution is to artificially increase by one gutter the width
of the inside image, then crop to the correct container width.