Fix instagram webview crash on Android
Instagram webview crashed when openning a page containing at least
one lazyloaded iframe.
It appears that the bug is due to the use of a svg dataurl in the
iframe src attribute. In any case, an iframe src should always be
a valid non empty url (SEE https://www.w3.org/TR/2011/WD-html5-20110525/the-iframe-element.html#the-iframe-element)
and as the "svg-in-src-to-preserve-aspect-ratio" is useless with
an iframe, this commit removes the src attribute completely when
lazyloading.